requirements.txt
anyio==3.6.2
Cantera==2.6.0.post1
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
flake8==6.0.0
Flask==2.2.3
fonttools==4.39.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
numpy==1.24.2
packaging==23.0
Pillow==9.5.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
typing_extensions==4.5.0
Werkzeug==2.2.3
